+++ src/main/java/org/jsoup/internal/StringUtil.java
@@ -387,5 +387,4 @@ public final class StringUtil {
             },
             StringJoiner::complete);
     }
-
 }